requirements.txt
aiohttp
flask
characterai
discord.py==2.4.0
DateTime==5.5
fortnite-api==2.6.8
jokeapi==1.0.5
requests==2.31.0
simplejson==3.19.2
